Melhoria da lista de produtos
